refactor: delete, add comments
